Sending Position To Server

+++ PC2/server/src/main/java/com/concurrente/server/server/ClientData.java
@@ -4,14 +4,15 @@ public class ClientData {
     private int index;
     private int money;
     private int energy;
-    private int x;
-    private int y;
+    private float x;
+    private float y;
     public ClientData(int index){
         this.index = index;
     }
-    public void update(int x, int y){
+    public void update(float x, float y){
         this.x = x;
         this.y = y;
+        //System.out.println("Cliente " + index + " actualizado en " + x + ", " + y);
     }
 
 }
